requirements.txt
anthropic>=0.25.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0